Adding a license (MIT).
Going with MIT so far because, honestly, it's one of the few licenses
I've ever read and fully understood, and I wish no one on this planet to
experience the agony involved in reading and trying to comprehend
horrors as the GPLv2 and the like.

I may decide to relicense later revisions under a more copyleft license
if this project ever gains much impulse.