Initial addition of windows and views
LOTS of work to be done in this area. However, I feel the fundamentals
are laid out right.

On top of that, Session now supports selecting existing windows and
creating new ones through the command line. If an invalid window is
selected, a new one is created, although this won't necessarily have the
invalid ID. (count yourself lucky if it does)

Session will alert the invoker of the ID of the newly created window.